Don't overwrite superclass attributes.
Overwriting superclass attributes was actually intended behavior,
because it allowed properties in superclasses to be overwritten in
subclasses.  However, if you want to customize how a property is defined
in a superclass, you don't want your customization to be overridden in
subclasses.  To account for this, this commit changes the behavior of
autoprop to only overwrite properties that were created by autoprop
itself.