feat: implement caching policies
The initial caching implementation didn't turn out to be very useful,
because it was too hard to control exactly how the properties would be
cached.  To address this, I implemented a system that allows a number of
different caching algorithms (called "policies") to be set on a
per-class and per-property basis.

BREAKING CHANGE: Drop support for python2.